# monolith/static/swagger.json
{
  "consumes": [
    "application/json",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Documentation of auto trade API.\n\nSchemes:\nhttp\nhttps",
    "title": "infoblog.",
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
...
    "/api/1/statistics/bot": {
      "post": {
        "tags": [
          "statistics"
        ],
        "summary": "Получение статистики бота по идентификатору.",
        "operationId": "botStatisticRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/BotStatisticRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/botStatisticResponse"
          }
        }
      }
    },
    "/api/1/statistics/delete": {
      "post": {
        "tags": [
          "statistics"
        ],
        "summary": "Удаление статистики бота по идентификатору.",
        "operationId": "botStatisticDeleteRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/BotStatisticDeleteRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/botStatisticDeleteResponse"
          }
        }
      }
    },
    "/api/1/statistics/user": {
      "get": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "tags": [
          "statistics"
        ],
        "summary": "Получение статистики всех ботов пользователя.",
        "operationId": "userStatisticRequest",
        "responses": {
          "200": {
            "$ref": "#/responses/userStatisticResponse"
          }
        }
      }
    },
    "/api/1/strategy/create": {
      "post": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "tags": [
          "strategy"
        ],
        "summary": "Создание стратегии.",
        "operationId": "strategyCreateRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/StrategyCreateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/strategyCreateResponse"
          },
          "400": {
            "description": " Bad request"
          },
          "500": {
            "description": " Internal server error"
          }
        }
      }
    },
...
    "/api/1/strategy/update": {
      "post": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "tags": [
          "strategy"
        ],
        "summary": "Редактирование стратегии.",
        "operationId": "strategyUpdateRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/StrategyUpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/strategyUpdateResponse"
          },
          "400": {
            "description": " Bad request"
          },
          "500": {
            "description": " Internal server error"
          }
        }
      }
    },
...
    "/api/1/user/changePassword": {
      "post": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "tags": [
          "user"
        ],
        "summary": "Изменение пароля пользователя.",
        "operationId": "changePasswordRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ChangePasswordIn"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/changePasswordResponse"
          },
          "400": {
            "$ref": "#/responses/changePasswordErrorResponse"
          }
        }
      }
    },
    "/api/1/user/profile": {
      "get": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "tags": [
          "user"
        ],
        "summary": "Получение профиля пользователя.",
        "operationId": "profileRequest",
        "responses": {
          "200": {
            "$ref": "#/responses/profileResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Bot": {
      "description": "go:generate easytags $GOFILE json,mapper",
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean",
          "x-go-name": "Active"
        },
        "asset_type": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "AssetType"
        },
        "auto_limit_sell_percent": {
          "type": "number",
          "format": "double",
          "x-go-name": "AutoLimitSellPercent"
        },
        "auto_sell": {
          "type": "boolean",
          "x-go-name": "AutoSell"
        },
        "commission_percent": {
          "type": "number",
          "format": "double",
          "x-go-name": "CommissionPercent"
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt"
        },
        "deleted_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "DeletedAt"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "exchange_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ExchangeID"
        },
        "exchange_type": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ExchangeType"
        },
        "exchange_user_key_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ExchangeUserKeyID"
        },
        "fixed_amount": {
          "type": "number",
          "format": "double",
          "x-go-name": "FixedAmount"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "kind": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Kind"
        },
        "limit_buy_percent": {
          "type": "number",
          "format": "double",
          "x-go-name": "LimitBuyPercent"
        },
        "limit_order": {
          "type": "boolean",
          "x-go-name": "LimitOrder"
        },
        "limit_sell_percent": {
          "type": "number",
          "format": "double",
          "x-go-name": "LimitSellPercent"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "order_count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "OrderCount"
        },
        "order_count_limit": {
          "type": "boolean",
          "x-go-name": "OrderCountLimit"
        },
        "order_type": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "OrderType"
        },
        "pair_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "PairID"
        },
        "pairs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StrategyPair"
          },
          "x-go-name": "Pairs"
        },
        "sell_percent": {
          "type": "number",
          "format": "double",
          "x-go-name": "SellPercent"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "UpdatedAt"
        },
        "user_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "UserID"
        },
        "uuid": {
          "type": "string",
          "x-go-name": "UUID"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/models"
    },
...
    "BotStatisticDeleteRequest": {
      "type": "object",
      "properties": {
        "bot_uuid": {
          "type": "string",
          "x-go-name": "BotUUID"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/statistics/controller"
    },
    "BotStatisticDeleteResponse": {
      "type": "object",
      "properties": {
        "error_code": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ErrorCode"
        },
        "success": {
          "type": "boolean",
          "x-go-name": "Success"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/statistics/controller"
    },
    "BotStatisticRequest": {
      "description": "go:generate easytags $GOFILE json",
      "type": "object",
      "properties": {
        "bot_uuid": {
          "type": "string",
          "x-go-name": "BotUUID"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/statistics/controller"
    },
    "BotStatisticResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/BotStatistics"
        },
        "error_code": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ErrorCode"
        },
        "success": {
          "type": "boolean",
          "x-go-name": "Success"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/statistics/controller"
    },
    "BotStatistics": {
      "description": "go:generate easytags $GOFILE json,mapper",
      "type": "object",
      "properties": {
        "bot_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "bot_uuid": {
          "type": "string",
          "x-go-name": "BotUUID"
        },
        "order_count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "OrderCount"
        },
        "profitability": {
          "type": "string",
          "x-go-name": "Profitability"
        },
        "user_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "UserID"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/models"
    },
...
    "ChangePasswordIn": {
      "type": "object",
      "properties": {
        "confirm_new_password": {
          "type": "string",
          "x-go-name": "ConfirmNewPassword"
        },
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "new_password": {
          "type": "string",
          "x-go-name": "NewPassword"
        },
        "old_password": {
          "type": "string",
          "x-go-name": "OldPassword"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/user/service"
    },
    "ChangePasswordOut": {
      "type": "object",
      "properties": {
        "error_code": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ErrorCode"
        },
        "success": {
          "type": "boolean",
          "x-go-name": "Success"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/user/service"
    },
    "Data": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/user/controller"
    },
...
    "ProfileResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/Data"
        },
        "error_code": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ErrorCode"
        },
        "success": {
          "type": "boolean",
          "x-go-name": "Success"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/user/controller"
    },
...
    "Strategy": {
      "description": "go:generate easytags $GOFILE json,mapper",
      "type": "object",
      "properties": {
        "bots": {
          "description": "UserID      int    `json:\"user_id\" mapper:\"user_id\"`\nAPIKey      string `json:\"api_key\" mapper:\"api_key\"`  TODO: нужно ли?\nSecretKey   string `json:\"secret_key\" mapper:\"secret_key\"`",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Bot"
          },
          "x-go-name": "Bots"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "exchange_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ExchangeID"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "uuid": {
          "type": "string",
          "x-go-name": "UUID"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/models"
    },
    "StrategyCreateRequest": {
      "type": "object",
      "properties": {
        "bots": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Bot"
          },
          "x-go-name": "Bots"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "exchange_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ExchangeID"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "uuid": {
          "type": "string",
          "x-go-name": "UUID"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/strategy/controller"
    },
    "StrategyCreateResponse": {
      "type": "object",
      "properties": {
        "error_code": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ErrorCode"
        },
        "strategy_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StrategyID"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/strategy/controller"
    },
...
    "StrategyPair": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "pair_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "PairID"
        },
        "strategy_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StrategyID"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/models"
    },
...
    "StrategyUpdateRequest": {
      "type": "object",
      "properties": {
        "fields": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int64"
          },
          "x-go-name": "Fields"
        },
        "strategy": {
          "$ref": "#/definitions/Strategy"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/strategy/controller"
    },
...
    "User": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "email_verified": {
          "type": "boolean",
          "x-go-name": "EmailVerified"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "phone": {
          "type": "string",
          "x-go-name": "Phone"
        },
        "phone_verified": {
          "type": "boolean",
          "x-go-name": "PhoneVerified"
        },
        "role": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Role"
        },
        "status": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Status"
        },
        "verified": {
          "type": "boolean",
          "x-go-name": "Verified"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/models"
    },
    "UserStatisticResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BotStatistics"
          },
          "x-go-name": "Data"
        },
        "error_code": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ErrorCode"
        },
        "success": {
          "type": "boolean",
          "x-go-name": "Success"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/statistics/controller"
    },
...
  },
  "responses": {
...
    "botStatisticDeleteResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/BotStatisticDeleteResponse"
      }
    },
    "botStatisticResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/BotStatisticResponse"
      }
    },
...
    "changePasswordErrorResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/ChangePasswordOut"
      }
    },
    "changePasswordResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/ChangePasswordOut"
      }
    },
...
    "profileResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/ProfileResponse"
      }
    },
...
    "strategyCreateResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/StrategyCreateResponse"
      }
    },
...
    "userStatisticResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/UserStatisticResponse"
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}